Added search feature

+++ app/views/posts/show.blade.php
@@ -13,9 +13,12 @@
 	</article>
 	
 	
-	<a class="btn btn-sm" href="{{{ action('PostsController@edit', $post->id) }}}">Edit</a>
-	{{ Form::open(['method' => 'DELETE', 'action' => ['PostsController@destroy', $post->id]]) }}
-		<button type="submit" class="btn btn-sm btn-danger">Delete</button>
-	{{ Form::close() }}
+	@if(Auth::check() && Auth::id() == $post->user_id)
+		<a class="btn btn-sm btn-primary pull-left" href="{{{ action('PostsController@edit', $post->id) }}}">Edit</a>
+		
+		{{ Form::open(['method' => 'DELETE', 'action' => ['PostsController@destroy', $post->id]]) }}
+			<button type="submit" class="btn btn-sm btn-danger">Delete</button>
+		{{ Form::close() }}
+	@endif
 	<script type="text/javascript" src="js/delete_confirm.js"></script>
 @stop
